requirements.txt
appdirs==1.4.4
asn1crypto==0.24.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
cachetools==3.1.1
certifi==2020.6.20
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
emoji==0.5.4
enum34==1.1.6
filelock==3.0.12
futures==3.3.0
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.6
importlib-metadata==1.6.1
importlib-resources==2.0.0
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
matplotlib==2.2.5
numpy==1.16.6
pathlib2==2.3.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pygobject==3.26.1
pyOpenSSL==17.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2020.1
pyxdg==0.25
requests==2.24.0
rsa==4.5
scandir==1.10.0
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.15.0
subprocess32==3.5.4
telepot==12.7
tornado==5.1.1
typing==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.23
zipp==1.2.0
